ai_service_backend/requirements.txt
fastapi
uvicorn
pdfplumber
python-docx
spacy
pydantic
